COMPONENTS FIELDS SELECT
Created the component that mounts the select view.
This component receives two parameters, one is a json with the information that will contain it and the other is the HTML where it will be inputted.
This component has validation whether it is mandatory or not.

The tests created for it are:
- should FieldsSelect exit
- Should FieldsSelect is a function
- Should create and append the markup give a correct data1 and markup1
- Should create and append the markup give a correct data2 and markup2