FEATURE TEXTAREA / INPUT
Created the TextareaFields component, this component puts the html textearea view in the view, this component receives two parameters. A date with the information that will fill it and an element that will be the place where it will be placed.

Created the InputFields component, this component puts the input html in the view, this component receives three sessions. A date with the information that will fill it, an element that will be the place where it is going to be placed and a type this type is the type of the input eg: text, email, tel ....

Tests created for TextareaFields:
- Should TextareaFields exit
- Should TextareaFields is a function
- should create and append the markup give a correct date and markup
- Must create and attach the markup without passing the placeholder to provide correct date and markup
- You must create and attach the markup as a required field to provide correct date and markup

Tests created for InputFields:
- Should TextareaFields exit
- Should TextareaFields is a function
- Should return an input of type cep since past date is type cep
- Should return an input of type text since past date is type text
- Should return an input of type email since last date is type email
- Should return to tel entry with input without being required
- Should return an input of type text if the type is not passed